BBB support. Updated config page to split RPi and BBB logic. Don't install RPi.GPIO on a BBB. Added BBB PWM pin setting.

+++ scripts/fpp_install.sh
@@ -3,13 +3,12 @@
 echo Installing python3-smbus...
 sudo apt-get install -y python3-smbus
 
-#echo Installing RPIO...
-#sudo pip install -U RPIO
-
-echo Installing RPi.GPIO
-sudo apt-get install -y python3-rpi.gpio
+if test -f /boot/config.txt; then
+  echo Installing RPi.GPIO
+  sudo apt-get install -y python3-rpi.gpio
+fi
 
 echo Restarting FPP...
 curl -s http://localhost/api/system/fppd/restart
-echo ...Done
+echo -e "\n...Done"
